gcp/fetilizer/requirements.txt
# matplotlib==3.8.0
seaborn==0.13.0
pandas==2.1.1
tensorflow==2.12.0
google-cloud-storage==1.16.1
# Pillow==6.0.0
scikit-learn==1.2.2

